requirements.txt
fastapi
uvicorn
pydantic
pyjwt
typing
bcrypt
psycopg2
